(TK-171) Add register-status protocol and function
Add a `register-status` function to the status service. This function can be
called by other trapperkeeper services that want their status reported by the
status service. The `register-status` function stores the service's name,
version, status callback function, and version of the status callback function
in an atom in the status service context.