(TK-171) Use pmap when calling status functions
pmap will be faster so long as more than one service is registered and at
least one of the `status-fn`s is somewhat computationally intensive. When this
is not the case, it is not vastly slower than `map` or `mapvals`. Thus, `pmap`
provides a significant performance improvement in the slow cases, and in
comparison is negligibly less performant when `status-fn` is fast.