(TK-174) Use handler/api rather than handler/site
handler/site from Compojure adds a lot of middleware that we don't need.
handler/api just adds middleware related to adding the query params into the
request map in a more useful way. This commit updates the routes to use
handler/api instead of handler/site, so we aren't using middleware that we
don't need.

Also, take advantage of the fact that handler/api uses `wrap-keyword-params`
and thus that params in `:params` are keywords (whereas params in
`:query-params` are not).