(maint) Pass status functions to routes, not atom
The `status-fns` atom in the status service's context should be updated uring
that service's init. Thus, by the time start is called it should be immutable.
This commit moves the dereferencing of the atom into `start` in the service
namespace, so that the route-related functions in the core namespace only deal
with a map.