(TK-209) Add :is-running to status service output
This commit adds an `:is-running` key to the status service output. It gets
the value for this key from the output of the status callback function that a
service registers. If the callback function does not return a map with an
:is-running key, or if the value for this key is not a boolean, the status
service returns :unknown.

This commit also adds a schema for the status callback response. The status
service does not actually validate that the result of calling the callback
function for a service matches this schema, but this schema can be used by a
service registering a callback function to validate the functions output.