(TK-210) Add utility function for getting service version
Because `service-version` is a required field for registering
a callback, and because we don't want to force all of the
consumers of this service to maintain a hard-coded instance
of that value by hand every time they do a new release of their
code, this commit introduces a utility function called
`get-service-version`.  If passed a valid maven group id and
artifact id, it will get the version string for that component.

This is intended to be used by downstream services; they should
call this prior to registering their callback, and then pass in
this value as the `service-version`.