(maint) Improve error response
Exceptions for pmaps get wrapped in `java.util.concurrent.ExecutionException`
since the exception happens in a future. However, this meant that if a schema
error happened when calling one of the status callback functions, it would get
wrapped by this and our `wrap-schema-error` middleware wouldn't catch it. This
commit unwraps the `ExecutionException` and rethrows the original exception.
It also pulls some additional keys out of the schema exception so that we get
a more useful error message in the case of an unexpected schema error.