(TK-209) Make value for :is-running a keyword
Add an `IsRunning` schema that can be reused in multiple places. This schema
changes the value of `:is-running` to always be a keyword: :true, :false, or
:unknown.

Add a TODO for merging the `StatusCallbackResponse` schema into the `Status`
schema, rather than duplicating its two keys. In rder to do this, we need to
make some changes to how we change hyphens to underscores for json output, but
for now leaving that for another time.