Add status.check namespace with disk & db check.
Add the `puppetlabs.trapperkeeper.services.status.check` namespace,
which contains shared status check functions for e.g. the database or
the disk.

Add `db-up?`, which takes a database connection spec and runs a purely
arithmetic query (which will succeed no matter the schema) and checks
the result; if it matches the expected result, the check returns `true`.
If the result doesn't match, or any exceptions are thrown in this
process, the check returns `false`

Add `disk-writable?`, which is a zero-arity function that attempts to
write random characters to a temporary file, reads the file back in, and
check that what was read matches what was written. If so, the check
returns `true`; if the match fails or any exceptions were thrown in this
process, the check returns `false`. The temporary file is deleted with
a `finally` clause.