(TK-231) Return 200 resp iff all services are up.
Change the status service's handler to return a 200 OK HTTP response if
and only if all the services being queried have the :is-running field of
their status set to :true. If any service has any other value for this
field, or is missing the field from that response entirely, return a 503
Service Unavailable response.