(TK-175) Add error handling for status check functions
This commit adds special handling around the calling of status check
functions. Specifically, it handles these cases:

 * Status check function times out
 * Status check function throws an Exception
 * Status check function returns something that doesn't conform to the
   callback response schema

In each error case, is_running is set to unknown and a descriptive
status key is set.

Note that none of these cases are reflected in the eventual TK status
service HTTP responses as that was not part of the specification of this
ticket.