(MAINT) Clean up and update deps, fix race
This commit cleans up the dependency list in project.clj, and
updates several of the dependencies to the latest versions.

Updating the trapperkeeper dependency changed some of the
behavior of the logging testutils, which revealed a race condition
in one of the tests.  This commit also fixes that race condition
and adds a bit more logic to the handling of the timeout
behavior when calling service status functions.