(MAINT) add exclusion for obsolete servlet-api
ring-defaults has a dependency on an old version of the javax
servlet API.  This artifact has a different name than new versions
of the servlet API, which means that lein won't detect a conflict,
and you will end up with a classpath that has multiple copies of
the servlet API in it, which will cause all kinds of indecipherable
errors, and that ain't no good time for nobody.

This commit adds an exclusion to prevent that terrible old dependency
from being brought in transitively by consumers of tk-status.