(TK-335) Make service lifecycle idempotent
This commit allows the status service to be stopped and started. Prior
to this commit, the status service's `start` function would attempt to
re-add the status service's own function to `:status-fns`, resulting in
a validation error.

Note that using TK's restart function formerly, and still, works fine as
it clears service context.