(TK-337) Fix race in timeout test
Previously, our status callback timeout test had a race condition
that caused it to almost always fail in Travis. This was due to
the test relying on an InterruptedException being thrown when
the status callback timed out and had its future cancelled.
However, due to the way Java was scheduling threads, that
was not always being thrown, so the assertion would frequently
fail.

This commit removes the frequently-failing test assertion, and
replaces it with a new assertion that tests the error message
that will be logged by the `with-timeout` macro, rather than
the one that should be logged if the timeout thread gets
interrupted.