(PE-16504) Add query parameter for timeout.
Change all of the status endpoints to accept a "timeout" query
parameter, which should be the timeout value in integral seconds.
If the timeout value is not provided, the default value will be whatever
`status-core/check-timeout` returns for the check's level. Any timeout
value that's not a positive integer will cause a 400 to be returned to
the client.