(TK-414) Add cpu usage metrics for app & GC
This commit introduces CPU usage metrics into the `:jvm-metrics` section
of the status callback.  It includes two fields, `:cpu-usage` and `:gc-cpu-usage`,
whose values are percentages of CPU capacity on the machine.

By default, we update the values of these metrics every 5 seconds.  The interval
can be modified by providing a value for the new setting
`[:status :cpu-metrics-interval-seconds]` in the trapperkeeper config.