(TK-455) Work around travis lein issues when backporting
During 0.6.0 travis supported a value of `lein2` in the `lein` key
of .travis.yml. This has been removed in current versions of
travis-ci and we took up this change in the 0.7.x as part of
our work to move to clj-parent.

Instead of pulling back the entire commit that brings in clj-parent
this commit only updates the .travis.yml file as to match the
content of 0.7.